pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Academic Paper Multimodal RAG System"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "docling<3.0.0",
    "docling-core<3.0.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "google-cloud-aiplatform>=1.140.0",
    "google-generativeai>=0.8.6",
    "numpy>=1.26.4",
    "pandas>=2.3.3",
    "pillow>=10.4.0",
    "pydantic-settings>=2.13.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "structlog>=25.5.0",
    "tqdm>=4.67.3",
    "typer==0.22",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
rag = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
exclude = ["^src/ingestion/ocr_extract\\.py$"]
disable_error_code = ["import-untyped"]
